♻️ [desktop]: extract autostart logic into dedicated module
Move macOS login item management functions from lib.rs to a new
autostart.rs module for better code organization.
